Copy buffer after flood fill to avoid overwriting
Using a NumPy array as a buffer increases performance, but leaves you
open to bugs due to reusing the same buffer and overwriting it, which is
what was happening. 😅

There's a better-performing fix still: use a buffer as big as the image,
but advance the start index in the buffer as you finish each flood fill.
This guarantees optimal space and time requirements (no need for copies,
or a huge buffer that is mostly unused).